Separate db_access into separate module
