+++ src/db/README.db.md
@@ -1,0 +1,3 @@
+# ugbio_db
+
+This module includes scripts for accessing PAPYRUS db
